+++ src/Catalog/Product.php
@@ -86,6 +86,12 @@ final class Product
     /** @var string */
     private $seoDescription;
 
+    /** @var \DateTimeImmutable|null */
+    private $createdAt;
+
+    /** @var \DateTimeImmutable|null */
+    private $updatedAt;
+
     /**
      * @internal
      */
@@ -131,6 +137,8 @@ final class Product
         }, $data['attachments'] ?? []);
         $this->seoTitle = $data['seoData']['title'] ?? '';
         $this->seoDescription = $data['seoData']['description'] ?? '';
+        $this->createdAt = isset($data['createdAt']) ? \DateTimeImmutable::createFromFormat(\DateTime::RFC3339, $data['createdAt']) : null;
+        $this->updatedAt = isset($data['updatedAt']) ? \DateTimeImmutable::createFromFormat(\DateTime::RFC3339, $data['updatedAt']) : null;
     }
 
     public function getId(): string
@@ -355,5 +363,15 @@ final class Product
     public function getSeoDescription(): string
     {
         return $this->seoDescription;
+    }
+
+    public function getCreatedAt(): ?\DateTimeInterface
+    {
+        return $this->createdAt;
+    }
+
+    public function getUpdatedAt(): ?\DateTimeInterface
+    {
+        return $this->updatedAt;
     }
 }